[np-50011] feat: Update OpenSearch cluster config (#100)
Ticket: https://sikt.atlassian.net/browse/NP-50011

As the existing cluster is lost, this is a good opportunity to upgrade the underlying node architecture. The template as defined in code has been out of sync with real infra for a long time, so we do not know exactly what we had running in prod anyway.

This is intended to let us run two separate clusters, effectively splitting prod and non-prod. This should let us scale down the prod cluster, so that the total cost is somewhat reasonable still. We can adjust node sizes later as needed.

Changes:
- Adds deletion protection to the OpenSearch domain and cognito pool to avoid unintended deletes during CloudFormation hiccups.
- Sets the data node instance type to `r7g.large.search`. This is roughly half the size of the nodes we had running in production, but with much more modern architecture and a node type favouring memory (`r` instead of `m` class).
- Increases the config for "non-prod" clusters to match the production cluster. It is kept as a separate config so that we can later bump instance size just for the production cluster, but for now they are the same.
- Enables zone awareness so that our data nodes are split across available data centres (at no extra cost).